fix(forms): use string type for option value input
Values bound to `HTMLOptionElement.value` will be converted into a string. The JSDoc comment points this out with: "Tracks simple string values bound to the option element. For objects, use the `ngValue` input binding."
However, the type of `NgSelectOption.value` is `any`. Strictening this type would cause an compiler error here to alert the developer that they should be using `[ngValue]` instead.